+++ release/i18next.amd-1.6.2.js
@@ -1,4 +1,4 @@
-// i18next, v1.6.2pre
+// i18next, v1.6.2
 // Copyright (c)2013 Jan Mühlemann (jamuhl).
 // Distributed under MIT license
 // http://i18next.com
@@ -943,6 +943,13 @@
             , lngs = options.lng ? f.toLanguages(options.lng) : languages
             , ns = options.ns || o.ns.defaultNs;
     
+        // split ns and key
+        if (key.indexOf(o.nsseparator) > -1) {
+            var parts = key.split(o.nsseparator);
+            ns = parts[0];
+            key = parts[1];
+        }
+    
         if (found === undefined && o.sendMissing) {
             if (options.lng) {
                 sync.postMissing(lngs[0], ns, key, notFound, lngs);
@@ -1084,7 +1091,7 @@
             if (o.fallbackNS.length) {
     
                 for (var y = 0, lenY = o.fallbackNS.length; y < lenY; y++) {
-                    found = _translate(o.fallbackNS[y] + o.nsseparator + key, options);
+                    found = _find(o.fallbackNS[y] + o.nsseparator + key, options);
                     
                     if (found) {
                         /* compare value without namespace */
@@ -1095,7 +1102,7 @@
                     }
                 }
             } else {
-                found = _translate(key, options); // fallback to default NS
+                found = _find(key, options); // fallback to default NS
             }
         }