+++ spec/init/init.namespace.spec.js
@@ -89,7 +89,8 @@ describe('with namespace', function() {
 
     describe('and fallbacking to multiple set namespace', function() {
       var resStore = {
-        dev: { 
+        dev: {
+          'ns.common': {},
           'ns.special': { 'simple_dev': 'ok_from_dev' },
           'ns.fallback1': { 
             'simple_fallback': 'ok_from_fallback1',
@@ -118,6 +119,37 @@ describe('with namespace', function() {
         expect(i18n.t('ns.common:simple_fallback')).to.be('ok_from_fallback1'); /* first wins */
         expect(i18n.t('ns.common:simple_fallback1')).to.be('ok_from_fallback1');
         expect(i18n.t('ns.common:simple_fallback2')).to.be('ok_from_fallback2');
+      });
+
+      describe('and post missing', function() {
+
+        var spy; 
+
+        beforeEach(function(done) {
+          spy = sinon.spy(i18n.sync, 'postMissing');
+          i18n.init(i18n.functions.extend(opts, { 
+            fallbackNS: ['ns.fallback1', 'ns.fallback2'], 
+            resStore: resStore,
+            sendMissing: true,
+            ns: { namespaces: ['ns.common', 'ns.special', 'ns.fallback'], defaultNs: 'ns.special'} } ),
+            function(t) { 
+              t('ns.common:notExisting');
+              done(); 
+            });
+        });
+
+        afterEach(function() {
+          spy.restore();
+        });
+
+        it('it should post only to origin namespace', function() {
+          expect(spy.callCount).to.be(1);
+          expect(spy.args[0][0]).to.be('en-US');
+          expect(spy.args[0][1]).to.be('ns.common');
+          expect(spy.args[0][2]).to.be('notExisting');
+          expect(spy.args[0][3]).to.be('ns.common:notExisting');
+        });
+
       });
 
     });